requirements.txt
# PKM System — Python dependencies
# Install with: pip install -r requirements.txt

# Core
pyyaml>=6.0
requests>=2.31

# PDF extraction
pymupdf>=1.24

# Web/article cleanup
beautifulsoup4>=4.12
lxml>=5.0

# YouTube download + transcription
yt-dlp>=2024.1
openai-whisper>=20231117

# Merge detection (local embeddings)
sentence-transformers>=3.0

# CLI
click>=8.1
